+++ CHANGES.md
@@ -1,4 +1,5 @@
 #5.0.0.rc1 2015-??-??
+ - 2015-09-29 Fixed bug#[11515](http://bugs.otrs.org/show_bug.cgi?id=11515) - Ticket notifications sends the content of the previous notification on the same event.
  - 2015-09-29 Fixed bug#[11507](http://bugs.otrs.org/show_bug.cgi?id=11507) - No ticket event notifications with out of office enabled, but not within ooO timeframe..
  - 2015-09-29 Updated DejaVu fonts to version 2.35.
  - 2015-09-25 Stats::Dynamic::TicketList: Add 'NumberOfArticles' ticket attribute. Thanks to Elias Probst.